BC-1422 - fix tests
