+++ src/components/organisms/administration/ImportUsers.vue
@@ -236,6 +236,14 @@
 						source: $t("pages.administration.migration.ldapSource"),
 					})
 				}}
+				<br />
+				<v-icon color="secondary">{{ mdiFlagOutline }}</v-icon>
+				{{
+					$t("components.organisms.importUsers.legendFlag", {
+						instance: this.$theme.short_name,
+						source: $t("pages.administration.migration.ldapSource"),
+					})
+				}}
 			</p>
 			<v-divider></v-divider>
 			<br />
@@ -245,8 +253,8 @@
 
 <script>
 /* eslint-disable max-lines */
-import { schoolsModule, importUsersModule } from "@/store";
-import { MatchedBy } from "@/store/import-users";
+import { importUsersModule, schoolsModule } from "@/store";
+import { MatchedBy } from "@store/import-users";
 import vImportUsersMatchSearch from "@components/molecules/vImportUsersMatchSearch";
 import {
 	mdiAccountPlus,
