+++ src/components/templates/TasksDashboardMain.vue
@@ -71,7 +71,6 @@
 </template>
 
 <script>
-import { taskModule, finishedTaskModule } from "@/store";
 import DefaultWireframe from "@/components/templates/DefaultWireframe.vue";
 import vCustomAutocomplete from "@components/atoms/vCustomAutocomplete";
 import vCustomSwitch from "@components/atoms/vCustomSwitch";
@@ -102,45 +101,67 @@ export default {
 			tasksEmptyStateImage,
 		};
 	},
+	inject: ["taskModule", "finishedTaskModule"],
 	computed: {
-		hasTasks: () => taskModule.hasTasks,
-		openTasksForStudentIsEmpty: () => taskModule.openTasksForStudentIsEmpty,
-		openTasksForTeacherIsEmpty: () => taskModule.openTasksForTeacherIsEmpty,
-		completedTasksForStudentIsEmpty: () =>
-			taskModule.completedTasksForStudentIsEmpty,
-		draftsForTeacherIsEmpty: () => taskModule.draftsForTeacherIsEmpty,
-		tasksCountStudent: () => taskModule.getTasksCountPerCourseStudent,
-		tasksCountTeacher: () => taskModule.getTasksCountPerCourseForTeacher,
-		isSubstituteFilterEnabled: () => taskModule.isSubstituteFilterEnabled,
-		courseFilters: () => taskModule.getCourseFilters,
-		selectedCourseFilters: () => taskModule.getSelectedCourseFilters,
-		finishedTasksIsInitialized: () => finishedTaskModule.getIsInitialized,
+		hasTasks() {
+			return this.taskModule.hasTasks;
+		},
+		openTasksForStudentIsEmpty() {
+			return this.taskModule.openTasksForStudentIsEmpty;
+		},
+		openTasksForTeacherIsEmpty() {
+			return this.taskModule.openTasksForTeacherIsEmpty;
+		},
+		completedTasksForStudentIsEmpty() {
+			return this.taskModule.completedTasksForStudentIsEmpty;
+		},
+		draftsForTeacherIsEmpty() {
+			return this.taskModule.draftsForTeacherIsEmpty;
+		},
+		tasksCountStudent() {
+			return this.taskModule.getTasksCountPerCourseStudent;
+		},
+		tasksCountTeacher() {
+			return this.taskModule.getTasksCountPerCourseForTeacher;
+		},
+		isSubstituteFilterEnabled() {
+			return this.taskModule.isSubstituteFilterEnabled;
+		},
+		courseFilters() {
+			return this.taskModule.getCourseFilters;
+		},
+		selectedCourseFilters() {
+			return this.taskModule.getSelectedCourseFilters;
+		},
+		finishedTasksIsInitialized() {
+			return this.finishedTaskModule.getIsInitialized;
+		},
 		// TODO: split teacher and student sides
-		isStudent: function () {
+		isStudent() {
 			return this.role === "student";
 		},
-		isTeacher: function () {
+		isTeacher() {
 			return this.role === "teacher";
 		},
-		showCourseFilter: function () {
+		showCourseFilter() {
 			if (this.tab === 2) return false;
 
 			return this.hasTasks;
 		},
-		showSubstituteFilter: function () {
+		showSubstituteFilter() {
 			return this.isTeacher && this.tab !== 2;
 		},
-		tabOneIsEmpty: function () {
+		tabOneIsEmpty() {
 			return this.isStudent
 				? this.openTasksForStudentIsEmpty
 				: this.openTasksForTeacherIsEmpty;
 		},
-		tabTwoIsEmpty: function () {
+		tabTwoIsEmpty() {
 			return this.isStudent
 				? this.completedTasksForStudentIsEmpty
 				: this.draftsForTeacherIsEmpty;
 		},
-		isCourseFilterDisabled: function () {
+		isCourseFilterDisabled() {
 			if (this.selectedCourseFilters.length > 0) return false;
 
 			if (this.tab === 0) {
@@ -152,7 +173,7 @@ export default {
 
 			return false;
 		},
-		getSortedCoursesFilters: function () {
+		getSortedCoursesFilters() {
 			const filters = this.courseFilters.map((filter) => {
 				const count = this.getTaskCount(filter.value);
 				const name = filter.value || this.$t("pages.tasks.labels.noCourse");
@@ -166,7 +187,7 @@ export default {
 
 			return filters.sort((a, b) => (a.text < b.text ? -1 : 1));
 		},
-		tabOneHeader: function () {
+		tabOneHeader() {
 			const tabOne = { icon: "$taskOpenFilled" };
 
 			tabOne.title = this.isStudent
@@ -175,7 +196,7 @@ export default {
 
 			return tabOne;
 		},
-		tabTwoHeader: function () {
+		tabTwoHeader() {
 			const tabTwo = {};
 			if (this.isStudent) {
 				tabTwo.icon = "$taskDoneFilled";
@@ -205,7 +226,7 @@ export default {
 			}
 			return null;
 		},
-		tabThreeHeader: function () {
+		tabThreeHeader() {
 			const tabThree = {
 				icon: "$taskFinished",
 				title: this.$t("components.organisms.TasksDashboardMain.tab.finished"),
@@ -213,7 +234,7 @@ export default {
 
 			return tabThree;
 		},
-		emptyState: function () {
+		emptyState() {
 			const image = tasksEmptyStateImage;
 			let title = "";
 			let subtitle = undefined;
@@ -246,10 +267,10 @@ export default {
 	},
 	methods: {
 		setCourseFilters(courseNames) {
-			taskModule.setCourseFilters(courseNames);
+			this.taskModule.setCourseFilters(courseNames);
 		},
 		setSubstituteFilter(enabled) {
-			taskModule.setSubstituteFilter(enabled);
+			this.taskModule.setSubstituteFilter(enabled);
 		},
 		getTaskCount(courseName) {
 			if (this.tab === 0) {
@@ -267,7 +288,7 @@ export default {
 			// TODO - this only properly works, because we switch between clients when archiving a task and therefor trigger a full reload
 			// we should probably find a better solution :D
 			if (!this.finishedTasksIsInitialized) {
-				finishedTaskModule.fetchFinishedTasks();
+				this.finishedTaskModule.fetchFinishedTasks();
 			}
 		},
 	},
